requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
streamlit>=1.27.2
chromadb>=0.4.18
google-generativeai>=0.3.0
pytesseract>=0.3.10
Pillow>=10.0.1
opencv-python>=4.8.1.78
python-multipart>=0.0.6
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0